Adds http js functions management, adds admin token
