--- hooks/double.js
@@ -1,9 +1,0 @@
-// @name: DOUBLE
-// @min_args: 1
-// @max_args: 1
-// @description: Doubles a number
-// @example: DOUBLE(5) returns 10
-
-function execute(args) {
-    return args[0] * 2;
-}
