--- hooks/wrong_example.js
@@ -1,9 +1,0 @@
-// @name: TRIPLE
-// @min_args: 1
-// @max_args: 1
-// @description: Triples a number
-// @example: TRIPLE(5) returns 15
-
-function execute(args) {
-    return args[0] * 3;
-}
