+++ src/bin/http_server/mod.rs
@@ -1,0 +1,7 @@
+pub mod auth;
+pub mod daemon;
+pub mod eval;
+pub mod js_management;
+pub mod stats;
+pub mod types;
+pub mod utils;
